refactor: switch to combinators where possible; update rust docs
