+++ src/data/handler/live/mod.rs
@@ -1,2 +1,5 @@
+/// Live data handler for [`Candle`] based [`MarketEvent`]s.
 pub mod candle;
+
+/// Live data handler for [`Trade`] (tick-by-tick) based [`MarketEvent`]s.
 pub mod trade;
